decrating: absorb shipper-cargo into shipper::ops::cargo
The former `shipper-cargo` microcrate is pulled into the core `shipper`
crate as `crate::ops::cargo`, following the Phase-2 decrating plan
(`docs/decrating-plan.md` §3.2). The historical public path
`shipper::cargo` is preserved via `pub use crate::ops::cargo` for
backward compatibility with `shipper-cli` and external consumers.

Scope of changes:

- New module `crates/shipper/src/ops/cargo/` containing merged
  functionality of the absorbed crate PLUS the pre-existing
  `crates/shipper/src/cargo.rs` (the `micro-cargo`-off branch's local
  implementation). The merged module exposes `CargoOutput`,
  `cargo_publish` / dry-run variants, `load_metadata`,
  `WorkspaceMetadata`, `PackageInfo`, `is_valid_package_name`,
  `workspace_member_names`, `get_version`, `get_package_name`, plus a
  `pub use shipper_output_sanitizer::redact_sensitive` re-export. All
  tests (unit, proptest, snapshot) from both the absorbed crate and the
  in-tree duplicate are ported; snapshot filenames rewritten to match
  the new module path.
- `shipper::cargo` is now `pub use crate::ops::cargo;` — no more
  `#[cfg(feature = "micro-cargo")]` dispatch. The `micro-cargo` feature
  is removed from `crates/shipper/Cargo.toml` and
  `crates/shipper-cli/Cargo.toml`.
- `crates/shipper/src/engine/parallel/publish.rs` swaps
  `use shipper_cargo as cargo;` for `use crate::ops::cargo;`.
- `crates/shipper/src/plan/mod.rs::load_metadata` now delegates to
  `crate::ops::cargo::load_metadata`.
- `crates/shipper-cargo/` physically deleted; dropped from workspace
  members in the root `Cargo.toml` and from `[dependencies]` in
  `crates/shipper/Cargo.toml`.
- Engine-parallel CLAUDE.md updated to reflect the new internal import.
- New `crates/shipper/src/ops/cargo/CLAUDE.md` describing invariants and
  gotchas of the absorbed module, matching the convention used by
  `ops/auth`, `ops/lock`, `ops/process`.

Pre-flight (per R-PR-4): the only consumer of `shipper-cargo` on main
was `crates/shipper/Cargo.toml` itself; `fuzz/` only references the
distinct `shipper-cargo-failure` crate. After this change the
`shipper_cargo::` namespace is gone from the workspace entirely.

Verified locally:

- `cargo check --workspace --all-targets` — clean
- `cargo check --workspace --all-targets --all-features` — clean
- `cargo clippy --workspace --all-targets --all-features -- -D warnings`
  — clean
- `cargo test --workspace` — 2000+ tests pass. One pre-existing
  `preflight_allow_dirty_snapshot` flake in `shipper-cli` was confirmed
  to fail identically on `origin/main` without these changes, so it is
  not caused by the absorption.

Part of the decrating effort — unblocks the next absorption round now
that `shipper-cargo` is off the dependency list of every crate other
than `shipper`.
