--- crates/shipper-cargo/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "shipper-cargo"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-description = "Cargo workspace metadata for shipper"
-repository = "https://github.com/EffortlessMetrics/shipper"
-documentation = "https://docs.rs/shipper-cargo"
-keywords = ["cargo", "metadata", "workspace"]
-categories = ["development-tools"]
-
-[dependencies]
-anyhow = "1.0"
-serde.workspace = true
-serde_json = "1.0"
-cargo_metadata = "0.23"
-humantime = "2.3.0"
-shipper-output-sanitizer = { path = "../shipper-output-sanitizer" }
-
-[dev-dependencies]
-tempfile = "3"
-shipper-output-sanitizer = { path = "../shipper-output-sanitizer" }
-proptest.workspace = true
-insta = { workspace = true }
-temp-env = "0.3"
-serial_test = "3"
-
-[lints]
-workspace = true
