+++ crates/shipper-cli/Cargo.toml
@@ -32,7 +32,6 @@ workspace = true
 [features]
 default = ["micro-all"]
 micro-encrypt = ["shipper/micro-encrypt"]
-micro-cargo = ["shipper/micro-cargo"]
 micro-webhook = ["shipper/micro-webhook"]
 micro-types = ["shipper/micro-types"]
 micro-config = ["shipper/micro-config"]
