--- crates/shipper/src/cargo_micro.rs
@@ -1,4 +1,0 @@
-pub use shipper_cargo::{
-    CargoOutput, cargo_publish, cargo_publish_dry_run_package, cargo_publish_dry_run_workspace,
-    redact_sensitive,
-};
